% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ar1_cor}
\alias{ar1_cor}
\title{Generate covariance matrix for MVN data}
\usage{
ar1_cor(p, rho)
}
\arguments{
\item{p}{number of input nodes}

\item{rho}{correlation of AR(1) process}
}
\value{
covariance matrix
}
\description{
Generate covariance matrix for MVN data
}
