% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_symmetries}
\alias{compare_symmetries}
\title{Compare two weight vectors to match up symmetries}
\usage{
compare_symmetries(W_pred, W_true, p, q, input_only = FALSE)
}
\arguments{
\item{W_pred}{weight vector to change}

\item{W_true}{comparison weight vector}

\item{p}{number of input nodes}

\item{q}{number of hidden nodes}

\item{input_only}{only compare using input-to-hidden weights (useful if
hidden-to-output weights grow too large due to redundancy)}
}
\value{
matrix containing all weight symmetries for W
}
\description{
Compare two weight vectors to match up symmetries
}
