% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_symmetries}
\alias{generate_symmetries}
\title{Generate all weight symmetries}
\usage{
generate_symmetries(W, p, q)
}
\arguments{
\item{W}{weight vector}

\item{p}{number of input nodes}

\item{q}{number of hidden nodes}
}
\value{
matrix containing all weight symmetries for W
}
\description{
Generate all weight symmetries
}
