% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_sic.R
\name{nn_sic_twostep}
\alias{nn_sic_twostep}
\title{Neural network SIC with twostep option}
\usage{
nn_sic_twostep(
  X,
  y,
  q,
  method = c("single", "group", "twostep"),
  grouping = c("input", "hidden"),
  eps_start = 10,
  t = 100,
  r = 0.87,
  n_iter = 1,
  n_init = 1,
  tau = 0
)
}
\arguments{
\item{X}{Input data}

\item{y}{Response vector}

\item{q}{Number of hidden nodes}

\item{method}{One of c("single", "group", "twostep")}

\item{grouping}{Grouping for "group" method}

\item{eps_start}{Starting epsilon value for telescoping}

\item{t}{Number of epsilon values}

\item{r}{Rate of decay for epsilon}

\item{n_iter}{Number of iterations for nlm after first step}

\item{n_init}{Number of random initialisations for first step}

\item{tau}{Value to swap from group to single method in "twostep"}
}
\value{
Function value
}
\description{
Neural network SIC with twostep option
}
