% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi_g}
\alias{phi_g}
\title{Smooth approximation to L0 norm for groups}
\usage{
phi_g(x, eps = 0)
}
\arguments{
\item{x}{Value}

\item{eps}{Epsilon value (default = 0)}
}
\value{
Function value
}
\description{
Smooth approximation to L0 norm for groups
}
