% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_recover}
\alias{weight_recover}
\title{Compare two weight vectors to match up symmetries}
\usage{
weight_recover(W_pred, W_true, p, q)
}
\arguments{
\item{W_pred}{weight vector to change}

\item{W_true}{comparison weight vector}

\item{p}{number of input nodes}

\item{q}{number of hidden nodes}

\item{input_only}{only compare using input-to-hidden weights (useful if
hidden-to-output weights grow too large due to redundancy)}
}
\value{
matrix containing all weight symmetries for W
}
\description{
Compare two weight vectors to match up symmetries
}
