% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{nn_logl_sic}
\alias{nn_logl_sic}
\title{SIC log-likelihood}
\usage{
nn_logl_sic(theta_s, y, X, q, eps, tanh_param = FALSE, c = NULL)
}
\arguments{
\item{theta_s}{A vector of the neural network parameters and log(sigma)}

\item{y}{Response vector}

\item{X}{Input matrix}

\item{q}{Number of hidden nodes}

\item{eps}{Epsilon value}

\item{tanh_param}{Switch to use hyperbolic tangent parametrisation for
constrained optimization}

\item{c}{constraint for tanh parametrisation}
}
\value{
negative penalised log-likelihood value
}
\description{
SIC log-likelihood
}
